Added auto switch to break duration

+++ TodoListMVVM/TodoListMVVM/Model/User/UserEntity+CoreDataProperties.swift
@@ -20,8 +20,8 @@ extension UserEntity {
     @NSManaged public var taskOverdueLimit: Int16
     @NSManaged public var userName: String?
     @NSManaged public var timerDuration: Int32
-    @NSManaged public var timerBreakDuration: Int16
-    @NSManaged public var timerRounds: Int16
+    @NSManaged public var timerBreakDuration: Int32
+    @NSManaged public var timerRounds: Int32
     
     public var wUserName: String {
         userName ?? "UserName"
